Arrg!  Getting this to work on linux as a service was a pain.
But it *should be* working now. Will let it run for a while. Both start up with upstart as services.

The big problem was that making it a daemon SHOULD require you to use 'expect daemon'. But that very explicitly DOESN'T work!

This combo does:
    # No expect statement
    exec $PYTHON_HOME/bin/python hnmonitor.py --daemon

Why? Maybe python's daemonize functionality does something unexpected with forks. Whatever.

Another HUGE tip - this can fail without realizing it, and that could be the problem, not a upstart problem. So, look in /var/log/upstart/<scriptname>.log  That's the first line of defense!